Ben Chaco (14):
      Added branch and build name arguments to build scripts
      Changes to build scripts for parallella-linux-adi repo renaming
      Updated Jenkins configuration file
      Install e-clear-shm-table and e-trace-dump utilities. Cleanup stdout formatting
      Added test for EPIPHANY_BUILD_TOOLCHAIN environment variable to bypass building the epiphany toolchain
      Fixed test for PARALLELLA_LINUX_HOME in build script
      Merge branch '2014.08.rc' of https://github.com/adapteva/epiphany-sdk into 2014.08.rc
      Merge branch '2014.08.rc' into jeremy-sdk-script-patch
      Merge branch 'jeremy-sdk-script-patch' into 2014.08.rc
      Merge branch '2014.08.rc' of https://github.com/adapteva/epiphany-sdk into 2014.08.rc
      Merge branch '2014.08.rc' of https://github.com/adapteva/epiphany-sdk into 2014.08.rc
      Merge branch '2014.08.rc' of https://github.com/adapteva/epiphany-sdk into 2014.08.rc
      Changes epiphany build and install scripts to support the latest Epiphany toolchain build script.
      Ignore errors when cloning paraellella-linux repo.

Jeremy Bennett (18):
      Rewrite of the download script to deal properly with GCC components.
      Consistency correction.
      Complete rewrite of build infrastructure for toolchain.
      Fix minor script typo.
      Added support for Canadian Cross toolchains.
      Better test for previously successful configure.
      Fix trivial typo.
      Updates to support integrated repositories and Canadian Cross builds
      Make build script for SDK more consisteng with toolchain script.
      use software-cache versions of repos.
      Fix synonyms
      Some of the install directory naming and linking was wrong.
      tighten up test for symbolic linking and forcing clean build.
      Correct ARM architecture name.
      Add ncurses support for Canadian Cross building
      Tidy up building of toolchain.
      Merge branch '2014.08.rc' of github.com:adapteva/epiphany-sdk into 2014.08.rc
      Remove irrelevant RPATH functionality.

Ola Jeppsson (19):
      Fix Canadian cross compilation
      install-sdk.sh: Install 'libe-trace.so'
      build-toolchain.sh: Don't install programs from ncurses
      build-toolchain.sh: Add staging directory.
      build-toolchain.sh: Install ncurses to staging area
      build-toolchain.sh: Install ncurses to staging area
      Have debug and clean flags affect entire build
      Don't install sudo wrapper scripts
      Merge pull request #32 from adapteva/pr-no-sudo
      Add multicore simulator
      Merge pull request #33 from adapteva/pr-add-multicore-simulator
      install-sdk.sh: Don't export CROSS_COMPILE unless it's necessary
      Set MANPATH in setup.*sh
      Support building by release tag
      Create 2014.11 branch
      setup.csh: Fix undefined MANPATH warning
      get-versions.sh: Fix detached head checkout failures
      Install buildhost's toolchain inside buildroot when cross-compiling
      Epiphany SDK 2014.11-rc2

Simon Cook (1):
      Improve building a canadian cross compiler.

aolofsson (1):
      Integration branch 2014.08.rc

+++ download-toolchain.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# Copyright (C) 2013 Embecosm Limited
+# Copyright (C) 2013,2014 Embecosm Limited
 
 # Contributor Simon Cook <simon.cook@embecosm.com>
 
@@ -17,108 +17,438 @@
 # more details.
 
 # You should have received a copy of the GNU General Public License along
-# with this program.  If not, see <http://www.gnu.org/licenses/>.  
+# with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-# Function to either download a tool or clone a git repository, checking out
-# the relevant branch.
-download_tool() {
+# Usage:
+
+# ./download-toolchain.sh [--force | --no-force]
+#                         [--clone | --download]
+#                         [--release | --no-release]
+#                         [--infra-url <url> | --infra-us |
+#                          --infra-uk | --infra-jp]
+#                         [--gmp | --no-gmp]
+#                         [--mpfr | --no-mpfr]
+#                         [--mpc | --no-mpc]
+#                         [--isl | --no-isl]
+#                         [--cloog | --no-cloog]
+#                         [--ncurses | --no-ncurses]
+#                         [--help | -h]
+
+# --force | --no-force
+
+#     If --force is specified, attempt to remove the existing repository
+#     before cloning or downloading. If the removal fails, the old version
+#     will be used, but a warning printed.  Default --no-force.
+
+# --clone | --download
+
+#     If --clone is specified, attempt to clone the repository, otherwise if
+#     --download is specified, attempt to download a ZIP file of the
+#     repository.  Default --download.
+
+# --release | --no-release
+
+#     If --release is specified, use ${RELEASE_TAG} specified in
+#     define-release.sh for all package versions.
+
+# --infra-url <url>
+
+#     Set the URL of the GCC infrastructure downloads. Default
+#     http://www.netgull.com/gcc/infrastructure.
+
+# --infra-us
+# --infra-uk
+# --infra-jp
+
+#     Synonyms respectively for infrastructure URLs in the USA, UK and Japan
+
+# --gmp | --no-gmp
+# --mpfr | --no-mpfr
+# --mpc | --no-mpc
+# --isl | --no-isl
+# --cloog | --no-cloog
+# --ncurses | --no-ncurses
+
+#     Download, or (with the --no- prefix) don't download the corresponding GCC
+#     infrastructure component. The components are downloaded from the
+#     infrastructure URL, which may be changed by the --infra-url option. By
+#     default all components are downloaded.
+
+# --multicore-sim | --no-multicore-sim
+
+#     Download, or (with the --no- prefix) don't download the GDB based
+#     experimental multi-core simulator.
+#     The components are downloaded from the
+#     infrastructure URL, which may be changed by the --infra-url option. By
+#     default all components are downloaded.
+
+
+# --help | -h
+
+#     Print out a brief help about this script and return with an error code.
+
+# Note that earlier versions allowed specification of a location where the
+# tools were to be downloaded. However recent versions do not download the SDK
+# directory, so the script will only work if downloading is done in a fixed
+# location relative to the pre-existing SDK directory (which contains this
+# script).
+
+# The script returns 1 on failure and 0 on success. Failure to delete a
+# pre-existing version when specifying --force is not considered a failure.
+
+
+################################################################################
+#                                                                              #
+#			       Shell functions                                 #
+#                                                                              #
+################################################################################
+
+# Function to clone a git repository, checking out the relevant branch.
+
+# The cloned repository will name its remote "adapteva".
+
+# @param[in] $1  The tool to clone
+# @param[in] $2  The full repository URL
+# @param[in] $3  The branch to checkout
+
+# @return 0 on success, 1 on failure. Note that failure to remove components
+#         when --force is in action is not considered a failure.
+clone_tool() {
     tool=$1
-    user=$2
-    repo=$3
-    branch=$4
+    repo_url=$2
+    branch=$3
 
-    if [ ${clone} = "true" ]
+    # If old source exists, delete
+    if [ ${force} = "true" ]
     then
-	# If old source exists, delete
-	if [ ${force} = "true" ]
+	if ! rm -rf ${tool} >> ${log} 2>&1
 	then
-	    if ! rm -rf ${tool} >> ${log} 2>&1
-	    then
-		echo "Unable to delete old ${tool}" | tee -a ${log}
-		return 0
-	    fi	
+	    echo "Warning: Unable to delete old ${tool}" | tee -a ${log}
 	fi
+    fi
 
-	# Clone git repository if it does not already exist
-	if [ -e ${tool} ]
-	then
-	    echo "Skipping ${tool} already exists..." | tee -a ${log}
-	else
-	    echo "Cloning ${tool}..."
-	    if ! git clone -q -b ${branch} git://github.com/${user}/${repo} ${tool} \
-	        >> ${log} 2>&1
-	    then
-		echo "Unable to clone ${tool}" | tee -a ${log}
-		return 1
-	    fi
-	fi
-
+    # Clone git repository if it does not already exist
+    if [ -e ${tool} ]
+    then
+	echo "${tool} already cloned." | tee -a ${log}
     else
-	# If old source exists, delete
-	if [ ${force} = "true" ]
+	echo "Cloning ${tool}..."
+	if ! git clone -q -o adapteva -b ${branch} ${repo_url} ${tool} \
+	         >> ${log} 2>&1
 	then
-	    if ! rm -rf ${tool} ${branch}.zip >> ${log} 2>&1
-	    then
-		echo "Unable to delete old ${tool}" | tee -a ${log}
-		return 0
-	    fi	
-	fi
-
-    	# Download and unzip source if it does not already exist
-	if [ -e ${tool} ]
-	then
-	    echo "Skipping ${tool}, already exists..." | tee -a ${log}
-	else
-	    echo "Downloading ${tool}..."
-	    if ! wget https://github.com/${user}/${repo}/archive/${branch}.zip \
-		>> ${log} 2>&1
-	    then
-		echo "Unable to download ${tool}" | tee -a ${log}
-		return 1
-	    fi
-	    if ! unzip ${branch}.zip >> ${log} 2>&1
-	    then
-		echo "Unable to unzip ${tool}" | tee -a ${log}
-		return 1
-	    fi
-	    if ! mv ${repo}-${branch} ${tool} >> ${log} 2>&1
-	    then
-		echo "Unable to move ${tool}" | tee -a ${log}
-		return 1
-	    fi
-	    if ! rm ${branch}.zip >> ${log} 2>&1
-	    then
-		echo "Unable to remove zip file for ${tool}" | tee -a ${log}
-		return 1
-	    fi
+	    echo "ERROR: Unable to clone ${tool}" | tee -a ${log}
+	    return 1
 	fi
     fi
 }
 
-# Check for relative directory and makes it absolute
-absolutedir() {
-  case ${1} in
-    /*) echo "${1}" ;;
-    *)  echo "${PWD}/${1}";;
-  esac
+# Function to download a specific tool branch or version
+
+# @param[in] $1  The tool to download
+# @param[in] $2  URL of archive directory
+# @param[in] $3  Command to unpack the download
+# @parma[in] $4  Branch or version packed file (i.e. with suffix such as .zip)
+# @parma[in] $5  Unpacked main directory
+
+# @return 0 on success, 1 on failure. Note that failure to remove components
+#         when --force is in action is not considered a failure.
+download_tool() {
+    tool=$1
+    archive_url=$2
+    unpack_cmd=$3
+    branch=$4
+    unpacked_dir=$5
+
+    # If --force is in action and old source exists, attempt delete
+    if [ ${force} = "true" ]
+    then
+	if ! rm -rf ${tool} ${branch} >> ${log} 2>&1
+	then
+	    echo "Warning: Unable to delete old ${tool}" | tee -a ${log}
+	fi
+    fi
+
+    # Download and unzip source if it does not already exist
+    if [ -e ${tool} ]
+    then
+	echo "${tool} already downloaded." | tee -a ${log}
+    else
+	echo "Downloading ${tool}..."
+	if ! wget ${archive_url}/${branch} \
+	    >> ${log} 2>&1
+	then
+	    echo "ERROR: Unable to download ${tool}" | tee -a ${log}
+	    return 1
+	fi
+	if ! ${unpack_cmd} ${branch} >> ${log} 2>&1
+	then
+	    echo "ERROR: Unable to unpack ${tool}" | tee -a ${log}
+	    return 1
+	fi
+	# Only move if the unpacked_dir is different to the tool
+	if [ "x${unpacked_dir}" != "x${tool}" ]
+	then
+	    if ! mv ${unpacked_dir} ${tool} >> ${log} 2>&1
+	    then
+		echo "ERROR: Unable to move unpacked dir to ${tool}" \
+		    | tee -a ${log}
+		return 1
+	    fi
+	fi
+	if ! rm ${branch} >> ${log} 2>&1
+	then
+	    echo "ERROR: Unable to remove packed file for ${tool}" \
+		| tee -a ${log}
+	    return 1
+	fi
+    fi
 }
 
-# Parse arguments
+# Function to either download a tool or clone a git repository from GitHub,
+# checking out the relevant branch.
+
+# @param[in] $1 The tool to clone (will appear as a subdirectory of ${topdir}.
+# @param[in] $2 The GitHub repo (within the adapteva organization) to
+#               clone/download.
+# @param[in] $3 The branch to checkout/download.
+
+# @param[in] $4 The release prefix tag (if any) Needed when we checkout
+# multiple heads in one directory (epiphany-binutils-gdb)
+
+# @return  The result of the underlying call to clone or download a tool.
+github_tool () {
+    tool=$1
+    repo=$2
+    branch=$3
+    release_prefix=$4
+
+    # If we're doing release we should use the release tags for all packages
+    # and since we use multiple heads in epiphany-binutils-gdb we need separate
+    # tags for each.
+    if [ "x--release" = "x${do_release}" ]
+    then
+	branch=${release_prefix}${RELEASE_TAG}
+    fi
+
+    if [ ${clone} = "true" ]
+    then
+	clone_tool "${tool}" "git://github.com/adapteva/${repo}" "${branch}"
+    else
+	download_tool "${tool}" "https://github.com/adapteva/${repo}/archive" \
+	              "unzip" "${branch}.zip" "${repo}-${branch}"
+    fi
+}
+
+# Function that loops over all component versions and downloads them
+
+# @return 0 on success, anything else indicates failure
+download_components() {
+    # Clone repositories from GitHub
+    # TODO: ??? Move this to get-versions.sh
+    # TODO: Some components might be optional and unnecessary to download.
+
+    OLD_IFS=${IFS}
+    IFS="
+" # We only want the newline character
+
+    res="ok"
+    for line in `cat ${basedir}/sdk/toolchain-components | grep -v '^#' \
+		     | grep -v '^$'`
+    do
+	tool=`          echo ${line} | cut -d ':' -f 1`
+	branch=`        echo ${line} | cut -d ':' -f 2`
+	repo=`          echo ${line} | cut -d ':' -f 3`
+	release_prefix=`echo ${line} | cut -d ':' -f 4`
+
+	if ! github_tool ${tool} ${repo} ${branch} ${release_prefix}
+	then
+	    res="fail"
+	    break
+	fi
+    done
+
+
+    # Restore IFS before returning
+    IFS=${OLD_IFS}
+
+    [ "xok" = "x${res}" ]
+}
+
+
+# Function to download a GCC component
+
+# @param[in] $1 Component name
+# @param[in] $2 File name with version
+# @param[in] $3 Compressed file suffix
+
+# @return  The result of the underlying call to download a tool.
+gcc_component () {
+    tool=$1
+    file=$2
+    packed_name=${file}.$3
+
+    if ! mkdir -p gcc-infrastructure
+    then
+	echo "ERROR: Unable to create gcc-infrastructure directory" \
+	    | tee -a ${log}
+	exit 1
+    fi
+
+    cd gcc-infrastructure
+    download_tool "${tool}" "${infra_url}" "tar xf" "${packed_name}" "${file}"
+    cd ..
+}
+
+
+# Function to download any other component
+
+# @param[in] $1 Component name
+# @param[in] $2 File name with version
+# @param[in] $3 Compressed file suffix
+# @param[in] $4 URL for the download
+
+# @return  The result of the underlying call to download a tool.
+other_component () {
+    tool=$1
+    file=$2
+    packed_name=${file}.$3
+    url=$4
+
+    if ! mkdir -p gcc-infrastructure
+    then
+	echo "ERROR: Unable to create gcc-infrastructure directory" \
+	    | tee -a ${log}
+	exit 1
+    fi
+
+    cd gcc-infrastructure
+    download_tool "${tool}" "${url}" "tar xf" "${packed_name}" "${file}"
+    cd ..
+}
+
+
+# Function to check for relative directory and makes it absolute
+
+# @param[in] $1  The directory to make absolute if necessary.
+absolutedir() {
+    case ${1} in
+
+	/*)
+	    echo "${1}"
+	    ;;
+
+	*)
+	    echo "${PWD}/${1}"
+	    ;;
+    esac
+}
+
+
+################################################################################
+#                                                                              #
+#			       Parse arguments                                 #
+#                                                                              #
+################################################################################
+
+# Defaults
+
 force="false"
 clone="false"
-checkout="${PWD}"
+do_release="--no-release"
+infra_url="http://mirrors-uk.go-parts.com/gcc/infrastructure"
+do_gmp="--do-gmp"
+do_mpfr="--do-mpfr"
+do_mpc="--do-mpc"
+do_isl="--do-isl"
+do_cloog="--do-cloog"
+do_ncurses="--do-ncurses"
+do_multicore_sim="--multicore-sim"
+
 until
 opt=$1
 case ${opt} in
+
     --force)
-	force="true" ;;
+	force="true"
+	;;
+
+    --no-force)
+	force="false"
+	;;
+
     --clone)
-	clone="true" ;;
-    --location)
-	checkout=$(absolutedir "$2"); shift ;;
+	clone="true"
+	;;
+
+    --download)
+	clone="false"
+	;;
+
+    --release | --no-release)
+	do_release="$1"
+	;;
+
+    --infra-url)
+	shift
+	infra_url="$1"
+	;;
+
+    --infra-us)
+	infra_url="http://www.netgull.com/gcc/infrastructure"
+	;;
+
+    --infra-uk)
+	infra_url="http://mirrors-uk.go-parts.com/gcc/infrastructure"
+	;;
+
+    --infra-jp)
+	infra_url="http://ftp.tsukuba.wide.ad.jp/software/gcc/infrastructure"
+	;;
+
+    --gmp | --no-gmp)
+	do_gmp="$1"
+	;;
+
+    --mpfr | --no-mpfr)
+	do_mpfr="$1"
+	;;
+
+    --mpc | --no-mpc)
+	do_mpc="$1"
+	;;
+
+    --isl | --no-isl)
+	do_isl="$1"
+	;;
+
+    --cloog | --no-cloog)
+	do_cloog="$1"
+	;;
+
+    --ncurses | --no-ncurses)
+	do_ncurses="$1"
+	;;
+
+    --multicore-sim | --no-multicore-sim)
+	do_multicore_sim="$1"
+	;;
+
     ?*)
-	echo "Usage: $0 [--force] [--clone] [--location <source location>]"
-	exit 1 ;;
+	echo "Usage: ./download-toolchain [--force | --no-force]"
+	echo "                            [--clone | --download]"
+	echo "                            [--infra-url <url> | --infra-us |"
+	echo "                             --infra-uk | --infra-jp]"
+	echo "                            [--gmp | --no-gmp]"
+	echo "                            [--mpfr | --no-mpfr]"
+	echo "                            [--mpc | --no-mpc]"
+	echo "                            [--isl | --no-isl]"
+	echo "                            [--cloog | --no-cloog]"
+	echo "                            [--ncurses | --no-ncurses]"
+	echo "                            [--multicore-sim | --no-multicore-sim]"
+	echo "                            [--help | -h]"
+	exit 1
+	;;
+
     *)
 	;;
 esac
@@ -127,23 +457,77 @@ do
     shift
 done
 
-# Create and move to checkout location
-mkdir -p "${checkout}"
-cd "${checkout}"
+
+################################################################################
+#                                                                              #
+#			     Download everything                               #
+#                                                                              #
+################################################################################
+
+# Move to basedir location
+
+d=`dirname "$0"`
+basedir=`(cd "$d/.." && pwd)`
+if ! cd "${basedir}"
+then
+    echo "ERROR: Unable to change to base directory for downloads/clones"
+    exit 1
+fi
+
+# Set the release parameters
+. ${basedir}/sdk/define-release.sh
 
 # Set up a log file
-log="${PWD}/clone-$(date -u +%F-%H%M).log"
+log="${LOGDIR}/clone-$(date -u +%F-%H%M).log"
 rm -f "${log}"
 
-# Clone repositories
-download_tool gcc           adapteva epiphany-gcc      epiphany-gcc-4.8
-download_tool binutils      adapteva epiphany-binutils epiphany-binutils-2.23
-download_tool gdb           adapteva epiphany-gdb      epiphany-gdb-7.6
-download_tool newlib        adapteva epiphany-newlib   epiphany-newlib-1.20
-download_tool cgen          adapteva epiphany-cgen     epiphany-cgen-1.1
+echo "Logging to ${log}"
 
-# BSC - we no longer checkout the SDK or LIBS here
-#download_tool sdk           adapteva epiphany-sdk      master
-#download_tool epiphany-libs adapteva epiphany-libs     master
 
-echo "Download complete"
+# Download all components defined in 'component-versions'
+if ! download_components
+then
+    echo "ERROR: Failed to download components" | tee -a ${log}
+    exit 1
+fi
+
+# Download optional GCC components
+# TODO: We want to define these outside of the download script, either in
+# 'toolchain-components' or 'gcc-components'.
+if [ "${do_gmp}" = "--do-gmp" ]
+then
+    gcc_component "gmp" "gmp-4.3.2" "tar.bz2" || res="fail"
+fi
+
+if [ "${do_mpfr}" = "--do-mpfr" ]
+then
+    gcc_component "mpfr" "mpfr-2.4.2" "tar.bz2" || res="fail"
+fi
+
+if [ "${do_mpc}" = "--do-mpc" ]
+then
+    gcc_component "mpc" "mpc-0.8.1" "tar.gz" || res="fail"
+fi
+
+if [ "${do_isl}" = "--do-isl" ]
+then
+    gcc_component "isl" "isl-0.12.2" "tar.bz2" || res="fail"
+fi
+
+if [ "${do_cloog}" = "--do-cloog" ]
+then
+    gcc_component "cloog" "cloog-0.18.1" "tar.gz" || res="fail"
+fi
+
+if [ "${do_ncurses}" = "--do-ncurses" ]
+then
+    other_component "ncurses" "ncurses-5.9" "tar.gz" \
+	"http://ftp.gnu.org/pub/gnu/ncurses" || res="fail"
+fi
+
+if [ "${res}" = "ok" ]
+then
+    echo "Download complete"
+else
+    echo "Download incomplete - see log for failures"
+fi
